Cleanup unnecessary components and code.
