+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,6 @@
         android:theme="@style/Theme.NearSpace">
         <activity
             android:name=".presentation.MainActivity"
-            android:exported="true" />
-        <activity
-            android:name=".presentation.BottomActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
