+++ app/src/main/res/layout/bottom_sheet_layout.xml
@@ -11,7 +11,7 @@
         android:layout_height="wrap_content"
         app:behavior_draggable="true"
         app:behavior_hideable="true"
-        app:behavior_peekHeight="10dp"
+        app:behavior_peekHeight="20dp"
         app:layout_behavior="@string/bottom_sheet_behavior">
 
         <View
@@ -23,7 +23,7 @@
 
         <TextView
             android:id="@+id/bottom_sheet_description_header"
-            style="@style/Description"
+            style="@style/DescriptionBottomSheet"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             app:layout_constraintEnd_toEndOf="parent"
