--- app/src/main/res/menu/menu_bottom_bar.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/app_bar_fav"
-        android:icon="@drawable/ic_favourite_menu"
-        android:title="@string/favourite"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/app_bar_settings"
-        android:icon="@drawable/ic_settings_menu"
-        android:title="@string/settings"
-        app:showAsAction="ifRoom" />
-</menu>
