+++ app/src/main/res/values/colors.xml
@@ -20,4 +20,6 @@
     <color name="orange_600">#F4511E</color>
     <color name="orange_900">#BF360C</color>
 
+    <color name="black_73_alfa">#BB000000</color>
+
 </resources>
