+++ app/src/main/res/values/styles.xml
@@ -10,18 +10,18 @@
 
     <style name="Line">
         <item name="android:layout_width">25dp</item>
-        <item name="android:layout_height">1dp</item>
+        <item name="android:layout_height">3dp</item>
         <item name="android:layout_marginTop">5dp</item>
         <item name="android:background">@color/line_background</item>
     </style>
 
-    <style name="Description">
+    <style name="DescriptionBottomSheet">
         <item name="android:layout_marginTop">5dp</item>
         <item name="android:paddingStart">10dp</item>
         <item name="android:paddingEnd">10dp</item>
-        <item name="android:textColor">@color/black</item>
+        <item name="android:textColor">@color/white</item>
         <item name="android:textStyle">bold</item>
-        <item name="android:background">?attr/colorPrimary</item>
+        <item name="android:background">@color/black_73_alfa</item>
     </style>
 
     <style name="Photo">
